Move and rename enable_output_polling() for clarity
enable_output_polling() does what it says, but its purpose is to make the server call back into do_send() when the socket is ready to send on.  Therefore rename it to schedule_send().

Move it into the connection class where it can use class members and does not require any arguments.