Implement fading on an 8-bit scale
This patch implements a fading backend, plus the UI bits to allow for
timed fading.

The backend should support user-controlled rather than time-controlled
fading, but that has not been implemented yet, mainly because I haven't
got a good idea as to how that could best be done.

[benh: Fix code formatting; run make update-po; add log entry; enable the
fade button later]