Improve UI feedback
Previously, when we lost all our dvsink-files connections, dvswitch
would disable the record sensitive property, but leave it active. As
such, when the dvsink connection returned, it would look like we're
recording (if looking at the wrong place), while we really aren't.

So, make sure the button isn't active when we set it to not be
sensitive.