Fix liveMedia library ordering, as CMake doesn't always do it
Traditional C linkers only resolve symbol references from system
libraries in libraries that are listed after them.  CMake is supposed
to hide this wrinkle by repeating the library list, but for some
reason it isn't doing this consistently.  Construct the library list
so that dependencies are listed after their dependents.