Fix polling logic and handling of buffer size mismatches
We must start polling when DVVideoStreamFramer requests a new frame
and stop polling as soon as we pass data back.  Since we cannot
control the size or alignment of data that libraw1394 gives us, we
must also handle the case where it is larger than the framer's buffer
without truncating.  Add an 'overspill' buffer for that.