send tally information when requested
Use the SET_PARAMETER command to set the parameter 'TALLY' to the
requested state.

This requires the upstream latest version of the LiveMedia libraries,
since the required sendSetParameter() call isn't (yet) available in the
Debian-packaged version.

The RFC requires that servers ignore parameters that they don't know
about, so applying this patch *should* be safe, even if the RTSP server
we're talking to doesn't understand tally lights.