Work around i18n issues
In some locales (e.g., Dutch), some of the shortcuts for the stock items
(e.g., "K_nippen" (cut) and "Op_nemen" (record)) conflict where they
don't in English. Work around this by manually rebuilding the stock
button out of its components, so that we can haz the English label in
the .po file and translate it to something containing a non-conflicting
shortcut in i18n. Also, don't limit this to the buttons that just happen
to be problematic in Dutch, since I don't know what other shortcuts in
other languages end up being.