Try to enable slice-per-thread encoding in newer libavcodec
libavocdec 52.111 introduced support for frame-per-thread processing.
We want to keep doing slice-per-thread encoding, and it appears we
need to request this in a different way now.  Move the dirty details
into auto_codec.

This removes the fallback to single-threaded operation in case
avcodec_thread_init() fails.  This fallback doesn't seem to have been
necessary in practice.