backwards compatible update to ffmpeg-0.10 API
libavcodec.53 deprecates quite a few functions used by dvswitch.
This patch wraps avcodec_open2(), avcodec_alloc_context3()
and discards avcodec_init().

Note on "__attribute__((unused))" in src/avcodec_wrap.h:
this file is used in both c and C++ code, so the variable name
can not be omitted (as would be the preferrec C++ way).
Not marking it as unused will prevent building with -Werror.
Be aware that "__attribute__((unused))" is gnu-gcc specific,
eventually a x-platform variant will me needed.