Fix more compilation errors
Apparently API stability is for wussies.

the RTSPClientSession class was changed to no longer have the stuff that
we need for tally light operation; instead, the relevant things have
moved to a new class "RTSPClientConnection". The difference is that a
Session may persist over several Connections.

Update our code to match.

Compiles, but does not link yet; this is likely due to a
multiarch-related change in packaging for libLiveMedia.