Fix manual fade behaviour
Previously, as detected during debconf, enabling the "manual fade"
function even once during a dvswitch session, would from then on
automatically enable it again once a secondary feed was chosen, even if
the manual fade function was disabled in the interface.

This is wrong, confusing for the dvswitch operator, and fairly annoying
for people watching the streams, so fix that behaviour.