Support libav, too
av_frame_alloc() exists in ffmpeg, and exists on the libav master branch
too, but does not yet exist in any released version of libav.

Test whether the symbol exists; if it doesn't, simulate it using calloc.

HACK: the actual av_frame_alloc() uses refcounting to release the data,
but we can't easily simulate that; so define an av_frame_dealloc that
does nothing in the "we have an av_frame_alloc" case, but releases the
data again in the other case, and call that from appropriate
destructors. This is slightly ugly, but it should work.